[CSM Portal Microapp] Merge Support page into a single cases list
Replaces the type-tabbed recent-5-then-"View All" pattern with one
infinite-scrolled, most-recently-updated-first case list on the
Support page itself, with search, filters, and create-case inline.
Drops the now-unreachable multi-type (service request/security
report/engagement/announcement) tabs and the separate AllCasesPage.
